Wireframes methods moved to THREEUtil
