+++ examples/3DTilesWireframe.html
@@ -70,7 +70,7 @@
         for (const tilesManager of frame3DPlanar.layerManager.tilesManagers)
           tilesManager.addEventListener(
             udvizBrowser.TilesManager.EVENT_TILE_LOADED,
-            udvizBrowser.appendWireframeToTileset
+            udvizBrowser.THREEUtil.appendWireframeToTileset
           );
       });
     </script>
