+++ packages/browser/src/index.js
@@ -10,11 +10,7 @@ export * from './Component/GUI/GUI';
 
 export { TilesManager } from './Component/Itowns/3DTiles/TilesManager';
 
-export {
-  getTileFromMesh,
-  appendWireframeToTileset,
-  appendWireframeByBatchIDToTileset,
-} from './Component/Itowns/3DTiles/3DTilesUtils';
+export { getTileFromMesh } from './Component/Itowns/3DTiles/3DTilesUtils';
 
 import * as ExternalScriptTemplate from './Component/Game/External/ScriptTemplate/ScriptTemplate';
 export { ExternalScriptTemplate };
